An initial implementation of import()
