+++ cproto/files.h
@@ -1,0 +1,6 @@
+#ifndef clox_files_h
+#define clox_files_h
+
+char* readFile(const char* path);
+
+#endif
